portable on linux

+++ libft/core/ft_memmove.c
@@ -14,11 +14,11 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	size_t i;
+	int i;
 
 	if (dst == src || len == 0)
 		return (dst);
-	i = dst > src ? len - 1 : -1;
+	i = dst > src ? (int)len - 1 : -1;
 	if (dst > src)
 	{
 		while (i + 1 >= 1)
@@ -29,7 +29,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	}
 	else
 	{
-		while (++i < len)
+		while (++i < (int)len)
 			*((unsigned char *)(dst + i)) = *((unsigned char *)(src + i));
 	}
 	return (dst);
